feat(s): implement "x-mode"
Replaces "full-inclusive" mode by extending it in the backward direction
with different semantics (`<c-x>` adds a +2 offset then, and the motion
remains exclusive). In both cases, the edge of the operated area gets an
offset of +2. A mnemonic could be: "backward exclusive, forward
extended".

This is an undocumented feature at the moment, the API has not changed.