perf(sx): `highlight-unique-chars` is fast enough now
This turned out to be ridiculously simple in the end, should have tried
this earlier. Getting the relevant lines from the buffer, and iterating
on the strings is _two_ orders of magnitude faster than using `search()`
- no clue what the latter is doing under the hood that makes it so
fantastically slow, but surely I will have a look at the implementation
sometime.

`opts.highlight_unique_chars` is of course on by default now.