feat(sx)!: smart auto-jump to first match
BREAKING CHANGE: API changes and changed defaults in `opts`.

Auto-jump to first match is an implicit, per sublist decision now,
based on a dead simple heuristic: the probability of the user aiming for
the very first match lessens with the number of matches.

`opts.labels` has been divided into a `labels` and a `safe_labels` list,
where the latter is assumed to contain characters that are unlikely to
be used right after a jump. When the number of matches does not exceed
the number of safe labels (plus one), we will automatically jump to the
first match, otherwise stay in place and revert to the default list.

The flag `jump_to_first_match` has been removed; the corresponding
behaviours can be forced by leaving either `labels` or `safe_labels`
empty.