fix: temporary highlight after jumping to a unique char
Problem: `highlight-new-curpos-and-op-area` should ignore the hacks
inside `jump-to!*`. (In Operator-pending mode we might push the cursor
beyond the actual target position, merely to simulate
inclusive/exclusive operational behaviour.)

Solution: `jump-to!*` now returns the final (intended) cursor position,
including adjustments for t-mode or x-mode, but ignoring the
aforementioned, additional pushes. After the jump,
`highlight-new-curpos...` gets that value as an argument, instead of
getting the cursor position itself.